Created Initial Structure for Columns
This commit contains the initial structure for the columns that will be needed on the dashboard page.

Furthermore, a naive card is created which will be used to create/test the drag-and-drop functionality. In this commit, the drag-and-drop functionality is counted in the console when running the application.

The commit also includes some minor styling corrections.